added 4 features, removed bad test
